reducer, acrion creator slice.js 에 통합
