+++ src/reducer.js
@@ -1,11 +1,3 @@
-import { reducer } from './slice';
+import reducer from './slice';
 
 export default reducer;
-
-// function defaultReducer(state) {
-//   return state;
-// }
-
-// export default function reducer(state = initialState, action) {
-//   return (reducers[action.type] || defaultReducer)(state, action);
-// }
